+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Implement deleting failed files from server feature (UIDATIMP-72)
 * Handle file name issue with file upload definition (UIDATIMP-97)
 * Show error message on UI when there is not enough server space for uploaded file(s) on uploading jop profiles (UIDATIMP-86)
+* Create data import settings page's 3rd pane for File Extensions (UIDATIMP-47)
 
 ## [1.0.0](https://github.com/folio-org/ui-data-import/tree/v1.0.0) (2018-11-10)